EP-6: Shall Statements (#4)
* The first draft of the shall statements for the Easypark system. Goes over functional and non functional requirements

* Markdown version of shall statements

Adjusted the shall statements file so that it can be properly presented in Markdown format

* Improved Shall Statements

An updated list of shall statements. Added more NFRs, changed existing FRs and NFRs, added descriptions for FRs and NFRs and changed the formatting in the md file

* Improved Shall Statements 2

Implemented feedback from the last set of shall statements. Generally clarified and improved requirements and added more rationales.